refactor: split the incremental jump analysis into per-build files
Replace the ZK_EVM conditionals with the repo's .std.cs/.zkevm.cs
convention, behind a single EvmStack.IsJumpDestination seam.
