+++ src/Nethermind/Nethermind.Evm/EvmStack.cs
@@ -73,31 +73,6 @@ public ref partial struct EvmStack
     private readonly CodeInfo? _codeInfo;
     private long[]? _jumpDestinations;
 
-    /// <summary>The jump-destination bitmap of <see cref="Code"/>, resolved on the first in-range jump and kept in the frame.</summary>
-    /// <remarks>
-    /// Kept on the stack so a jump validates against the frame it is executing without walking
-    /// <c>vm.VmState.Env.CodeInfo</c>. Resolving lazily keeps the analysis off the path of frames that
-    /// never jump. Only a stack built over no code may omit the code info; the empty bitmap then rejects
-    /// every destination.
-    /// </remarks>
-    internal long[] JumpDestinations
-    {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        get
-        {
-            Debug.Assert(_codeInfo is not null || CodeLength == 0, "A stack that executes code must carry that code's CodeInfo.");
-#if ZK_EVM
-            return _jumpDestinations ??= _codeInfo?.IncrementalJumpBitmap ?? JumpDestinationAnalyzer.EmptyBitmap;
-#else
-            return _jumpDestinations ??= _codeInfo?.JumpDestinationBitmap ?? JumpDestinationAnalyzer.EmptyBitmap;
-#endif
-        }
-    }
-
-#if ZK_EVM
-    internal bool AnalyzeJump(int destination) => _codeInfo!.AnalyzeJump(destination);
-#endif
-
     /// <summary>
     /// Reserves the next stack slot and returns a ref to it. On overflow returns <see cref="Unsafe.NullRef{T}"/>;
     /// callers must check with <see cref="Unsafe.IsNullRef{T}"/> before writing.
