+++ src/Nethermind/Nethermind.Evm/EvmStack.std.cs
@@ -1,14 +1,32 @@
 // SPDX-FileCopyrightText: 2026 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Nethermind.Core.Extensions;
+using Nethermind.Evm.CodeAnalysis;
 using Nethermind.Int256;
 
 namespace Nethermind.Evm;
 
 public ref partial struct EvmStack
 {
+    /// <summary>Reports whether <paramref name="destination"/> is a valid jump destination in <see cref="Code"/>.</summary>
+    /// <remarks>
+    /// The bitmap is resolved on the first in-range jump and kept in the frame, so a jump validates
+    /// against the frame it is executing without walking <c>vm.VmState.Env.CodeInfo</c>, and a frame
+    /// that never jumps never pays for the analysis. Only a stack built over no code may omit the code
+    /// info; the empty bitmap then rejects every destination. See <c>EvmStack.zkevm.cs</c> for the guest
+    /// form, which analyzes the code only as far as it jumps into it.
+    /// </remarks>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    internal bool IsJumpDestination(int destination)
+    {
+        Debug.Assert(_codeInfo is not null || CodeLength == 0, "A stack that executes code must carry that code's CodeInfo.");
+        long[] bitmap = _jumpDestinations ??= _codeInfo?.JumpDestinationBitmap ?? JumpDestinationAnalyzer.EmptyBitmap;
+        return JumpDestinationAnalyzer.IsJumpDestination(bitmap, destination);
+    }
+
     /// <summary>Writes <paramref name="value"/> as one big-endian 32-byte stack word.</summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private static void WriteBeWord(ref EvmWord head, in UInt256 value)
