+++ src/Nethermind/Nethermind.Evm/EvmStack.zkevm.cs
@@ -1,14 +1,32 @@
 // SPDX-FileCopyrightText: 2026 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Nethermind.Core.Extensions;
+using Nethermind.Evm.CodeAnalysis;
 using Nethermind.Int256;
 
 namespace Nethermind.Evm;
 
 public ref partial struct EvmStack
 {
+    /// <summary>Reports whether <paramref name="destination"/> is a valid jump destination in <see cref="Code"/>.</summary>
+    /// <remarks>
+    /// The frame caches the code info's incremental bitmap, which covers only the code scanned so far,
+    /// so a clear bit is not yet an answer: it falls through to <see cref="CodeInfo.AnalyzeJump"/>, which
+    /// extends the scan to reach the destination. Repeat jumps to an already scanned destination take the
+    /// bit test alone. See <c>EvmStack.std.cs</c> for the host form, which analyzes the whole code up front.
+    /// </remarks>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    internal bool IsJumpDestination(int destination)
+    {
+        Debug.Assert(_codeInfo is not null || CodeLength == 0, "A stack that executes code must carry that code's CodeInfo.");
+        long[] bitmap = _jumpDestinations ??= _codeInfo?.IncrementalJumpBitmap ?? JumpDestinationAnalyzer.EmptyBitmap;
+        // The caller has already bounded the destination by the code length, so the code info is present.
+        return JumpDestinationAnalyzer.IsJumpDestination(bitmap, destination) || _codeInfo!.AnalyzeJump(destination);
+    }
+
     /// <summary>Writes <paramref name="value"/> as one big-endian 32-byte stack word.</summary>
     /// <remarks>
     /// Carries the same small-value shortcut as <see cref="ReadBeWord"/>: pushed values are
